refactor wordlist_skill by vocabulary, added module urls
