+++ Emmily/settings.py
@@ -31,8 +31,8 @@ INSTALLED_APPS = [
     'rest_framework.authtoken',
     'rest_framework',
     'django_filters',
-    'wordlist_skill',
-    'account'
+    'account',
+    'vocabulary'
 ]
 
 REST_FRAMEWORK = {
