+++ Emmily/urls.py
@@ -1,6 +1,7 @@
 from django.conf.urls import url
 from django.contrib import admin
 from django.urls import path
+from django.urls.conf import include
 
 from account.views import (
     registration_view,
@@ -11,17 +12,13 @@ from account.views import (
     ObtainAuthTokenView
 )
 
-import  wordlist_skill.api_views
-import  wordlist_skill.views
-
 urlpatterns = [
+    # REST-framwork
+    path('api/v1/vocabulary/', include('vocabulary.urls')),
+    
     path('admin/', admin.site.urls),
     path('account/', account_view, name="account"),
     path('api/v1/account/login', ObtainAuthTokenView.as_view(), name="login"),
-
-    path('api/v1/wordlist/', wordlist_skill.api_views.WordGetList.as_view()),
-    path('api/v1/wordlist/new', wordlist_skill.api_views.WordCreation.as_view()),
-    path('api/v1/wordlist/<int:id>', wordlist_skill.api_views.WordRetriveUpdateDestroy.as_view()),
-
-    url('',wordlist_skill.views.index)
+    
+    #url('',wordlist_skill.views.index)
 ]
