+++ vocabulary/api_views.py
@@ -3,10 +3,10 @@ from rest_framework.authentication import TokenAuthentication
 from django_filters.rest_framework import DjangoFilterBackend
 from rest_framework.pagination import LimitOffsetPagination
 from rest_framework.exceptions import ValidationError
-from wordlist_skill.serializers import WordListSerializer, WordSerializer
+from vocabulary.serializers import WordListSerializer, WordSerializer
 from rest_framework.permissions import IsAdminUser
 from rest_framework.filters import SearchFilter
-from wordlist_skill.models import Word, WordList, WordList
+from vocabulary.models import Word, WordList
 
 class Pagination(LimitOffsetPagination):
     default_limit = 10
@@ -45,7 +45,7 @@ class WordCreation(CreateAPIView):
         return super().create(request,*args,**kwargs)   
 
 class WordRetriveUpdateDestroy(RetrieveUpdateDestroyAPIView):
-    permission_classes = [IsAdminUser]   
+    permission_classes = [IsAdminUser]
     queryset = Word.objects.all()
     lookup_field = 'id'
     serializer_class = WordSerializer
@@ -77,62 +77,3 @@ class WordRetriveUpdateDestroy(RetrieveUpdateDestroyAPIView):
 #endregion
 
 
-#region WordList
-class WordListGetList(ListAPIView):
-    authentication_classes = [TokenAuthentication]
-    permission_classes = [IsAdminUser]
-    queryset = WordList.objects.all()
-    serializer_class = WordListSerializer
-    filter_backends = (DjangoFilterBackend,SearchFilter)
-    filter_fields = ('id',)
-    search_fields = ('name',"language","owner")
-    pagination_class = Pagination
-
-    def get_queryset(self):
-            #Filter by language
-            request_language = self.request.query_params.get('lan', None)
-            queryset = Word.objects.all()
-            if request_language is not None: 
-                queryset = queryset.filter(
-                    language=request_language
-                )
-            return queryset
-
-
-class WordCreation(CreateAPIView):
-    permission_classes = [IsAdminUser]
-    serializer_class = WordListSerializer
-    
-    def create(self,request,*args,**kwargs):
-        #Check if the word already exist
-        if WordList.objects.filter(name = request.data.get("name")).count() != 0:
-            raise ValidationError({'WordList name':'Already existe on the DB'})
-        return super().create(request,*args,**kwargs)  
-
-class WordListRetriveUpdateDestroy(RetrieveUpdateDestroyAPIView):
-    permission_classes = [IsAdminUser]   
-    queryset = Word.objects.all()
-    lookup_field = 'id'
-    serializer_class = WordListSerializer
-
-    def delete(self,request,*args, **kwargs):
-        #Clear cache
-        product_id = request.data.get('id')
-        response = super().delete(self,request,*args, **kwargs)
-        if response.status_code == 204:
-            from django.core.cache import cache
-            cache.delete('product_data_{}'.format(product_id))
-        return response
-
-    def update(self,request,*args, **kwargs):
-        response = super().update(request,*args, **kwargs)
-        if response.status_code == 200:
-            from django.core.cache import cache
-            word = response.data
-            cache.set('word_data_{}'.format(word['id']),{
-                'name': word['name'],
-                'language': word['language']
-            })
-        return response
-
-#endregion
