+++ vocabulary/models.py
