--- wordlist_skill/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class WordlistSkillConfig(AppConfig):
-    name = 'wordlist_skill'
