--- wordlist_skill/views.py
@@ -1,5 +1,0 @@
-from django.http import HttpResponse
-from django.shortcuts import render
-
-def index(request):
-    return render(request,'index.html')
